Refactor: move scheduler code to a separate folder
