+++ src/lib.rs
@@ -56,6 +56,7 @@ mod mmtk;
 mod plan;
 pub mod policy;
 pub mod vm;
+pub mod scheduler;
 
 pub use crate::mm::memory_manager;
 pub use crate::mmtk::MMTK;
@@ -65,6 +66,5 @@ pub use crate::plan::selected_plan::{
 pub use crate::plan::{
     Allocator, CollectorContext, Mutator, MutatorContext, ParallelCollector, Plan, TraceLocal,
     TransitiveClosure,
-    worker, scheduler, work,
     CopyContext,
 };
